refactor unit parsing to enable storing unit expressions in the unit system
