+++ unyt/unit_object.py
@@ -16,12 +16,9 @@ A class that represents a unit symbol.
 import copy
 import itertools
 import math
-
-from functools import lru_cache
-from keyword import iskeyword as _iskeyword
 import numpy as np
+from functools import lru_cache
 from numbers import Number as numeric_type
-import token
 
 from sympy import (
     Expr,
@@ -30,17 +27,14 @@ from sympy import (
     Number,
     Pow,
     Symbol,
-    Integer,
     Float,
     Basic,
     Rational,
-    sqrt,
     Mod,
     floor,
 )
 from sympy.core.numbers import One
 from sympy import sympify, latex
-from sympy.parsing.sympy_parser import parse_expr, auto_number, rationalize
 
 from unyt.dimensions import (
     angle,
@@ -58,92 +52,14 @@ from unyt.exceptions import (
     UnitConversionError,
     UnitDtypeError,
     UnitsNotReducible,
-)
-from unyt._physical_ratios import speed_of_light_cm_per_s
-from unyt._unit_lookup_table import inv_name_alternatives
-from unyt.unit_registry import (
-    default_unit_registry,
-    _lookup_unit_symbol,
-    UnitRegistry,
     UnitParseError,
 )
+from unyt._parsing import parse_unyt_expr
+from unyt._physical_ratios import speed_of_light_cm_per_s
+from unyt.unit_registry import default_unit_registry, _lookup_unit_symbol, UnitRegistry
 from unyt.unit_systems import _split_prefix
 
 sympy_one = sympify(1)
-
-global_dict = {
-    "Symbol": Symbol,
-    "Integer": Integer,
-    "Float": Float,
-    "Rational": Rational,
-    "sqrt": sqrt,
-}
-
-
-def _auto_positive_symbol(tokens, local_dict, global_dict):
-    """
-    Inserts calls to ``Symbol`` for undefined variables.
-    Passes in positive=True as a keyword argument.
-    Adapted from sympy.sympy.parsing.sympy_parser.auto_symbol
-    """
-    result = []
-    prevTok = (None, None)
-
-    tokens.append((None, None))  # so zip traverses all tokens
-    for tok, nextTok in zip(tokens, tokens[1:]):
-        tokNum, tokVal = tok
-        nextTokNum, nextTokVal = nextTok
-        if tokNum == token.NAME:
-            name = tokVal
-            if (
-                name in ["True", "False", "None"]
-                # special case 'as' for attosecond
-                or (_iskeyword(name) and name != "as")
-                or name in local_dict
-                # Don't convert attribute access
-                or (prevTok[0] == token.OP and prevTok[1] == ".")
-                # Don't convert keyword arguments
-                or (
-                    prevTok[0] == token.OP
-                    and prevTok[1] in ("(", ",")
-                    and nextTokNum == token.OP
-                    and nextTokVal == "="
-                )
-            ):
-                result.append((token.NAME, name))
-                continue
-            elif name in global_dict:
-                obj = global_dict[name]
-                if isinstance(obj, (Basic, type)) or callable(obj):
-                    result.append((token.NAME, name))
-                    continue
-
-            # try to resolve known alternative unit name
-            try:
-                used_name = inv_name_alternatives[str(name)]
-            except KeyError:
-                # if we don't know this name it's a user-defined unit name
-                # so we should create a new symbol for it
-                used_name = str(name)
-
-            result.extend(
-                [
-                    (token.NAME, "Symbol"),
-                    (token.OP, "("),
-                    (token.NAME, repr(used_name)),
-                    (token.OP, ","),
-                    (token.NAME, "positive"),
-                    (token.OP, "="),
-                    (token.NAME, "True"),
-                    (token.OP, ")"),
-                ]
-            )
-        else:
-            result.append((tokNum, tokVal))
-
-        prevTok = (tokNum, tokVal)
-
-    return result
 
 
 def _get_latex_representation(expr, registry):
@@ -192,9 +108,6 @@ def _get_latex_representation(expr, registry):
         return ""
     else:
         return latex_repr
-
-
-unit_text_transform = (_auto_positive_symbol, rationalize, auto_number)
 
 
 class Unit(object):
@@ -257,22 +170,7 @@ class Unit(object):
             if registry and unit_expr in registry._unit_object_cache:
                 return registry._unit_object_cache[unit_expr]
             unit_cache_key = unit_expr
-            if not unit_expr:
-                # Bug catch...
-                # if unit_expr is an empty string, parse_expr fails hard...
-                unit_expr = "1"
-            try:
-                unit_expr = parse_expr(
-                    unit_expr,
-                    global_dict=global_dict,
-                    transformations=unit_text_transform,
-                )
-            except SyntaxError as e:
-                msg = "Unit expression '%s' raised an error during parsing:\n%s" % (
-                    unit_expr,
-                    repr(e),
-                )
-                raise UnitParseError(msg)
+            unit_expr = parse_unyt_expr(unit_expr)
         # Simplest case. If user passes a Unit object, just use the expr.
         elif isinstance(unit_expr, Unit):
             # grab the unit object's sympy expression.
